+++ CHANGELOG.md
@@ -7,5 +7,6 @@ All notable changes made by us to the source code of the Anki-Connect project ar
 
 ## Additions
 - Return of **note modification time** in the `notesInfo` method
+- Added optional parameter `collapsedFields` in Method `createModel`
 - New method: `notesModTime` that retrieves modification timestamps for notes
 - New method: `isFsrsActive` that detects usage of FSRS for scheduling